Move get_time to new utils.time module
