+++ huracanpy/_tracker_specific/csv.py
@@ -6,34 +6,7 @@ import numpy as np
 import xarray as xr
 import pandas as pd
 
-def get_time(year, month, day, hour):
-    """
-    Get np.datetime64 array corresponding to year, month, day and hour arrays
-
-    Parameters
-    ----------
-    year (np.array or pd.Series)
-    month (np.array or pd.Series)
-    day (np.array or pd.Series)
-    hour (np.array or pd.Series)
-
-    Returns
-    -------
-    np.array or pd.Series
-        The corresponding np.datetime64
-    """
-    time = pd.to_datetime(
-        year.astype(str)
-        + "-"
-        + month.astype(str)
-        + "-"
-        + day.astype(str)
-        + " "
-        + hour.astype(str)
-        + ":00"
-    )
-    return time
-
+from .utils.time import get_time
 
 def load(filename,):
     """Load csv tracks data as an xarray.Dataset
